Adding list_cast to the casts module
This takes string representation of a list and returns python list.
The default behaviour to use a comma as the delimiter, but alternate delimiters can be specified as well.

If an optional base_cast function is supplied, it will be called on each member of the list.

A set of tests are added to test_casts.py to test all of the params.